requirements.txt
flask==3.0.0
gunicorn==21.2.0
pypdf==4.0.1
werkzeug==3.0.1
sqlalchemy==2.0.25
python-dotenv==1.0.0
requests==2.31.0
